+++ Apps/AquaScriptCore/Components/NativeComponent.cs
@@ -16,7 +16,7 @@ namespace AquaEngine
 			}
 		}
 
-		[MethodImpl(MethodImplOptions.InternalCall)] private static extern IntPtr _GetHandle(uint worldID, uint entityID, Type type);
+		[MethodImpl(MethodImplOptions.InternalCall)] private static extern IntPtr _GetHandle(ulong worldID, ulong entityID, Type type);
 
 	}
 }
