+++ Apps/AquaScriptCore/Systems/NativeSystem.cs
@@ -16,7 +16,7 @@ namespace AquaEngine
 			}
 		}
 
-		[MethodImpl(MethodImplOptions.InternalCall)] private static extern IntPtr _GetHandle(uint worldID, Type type);
+		[MethodImpl(MethodImplOptions.InternalCall)] private static extern IntPtr _GetHandle(ulong worldID, Type type);
 	}
 }
