+++ Engine/Source/Graphics/Shader.cpp
@@ -301,7 +301,7 @@ ShaderUniform Shader::GetUniformInfo(std::string locationName) const
 #include <AquaEngine/Resource.hpp>
 #include <AquaEngine/Scripting/InternalCalls.hpp>
 
-ADD_MANAGED_METHOD(Shader, Load0, void, (MonoString* path, unsigned int* outResourceID, void** outHandle), AquaEngine.Graphics)
+ADD_MANAGED_METHOD(Shader, Load0, void, (MonoString* path, uint64_t* outResourceID, void** outHandle), AquaEngine.Graphics)
 {
 	*outResourceID = Resource::Load<Shader>(mono_string_to_utf8(path));
 	*outHandle = Resource::Get<Shader>(*outResourceID);
@@ -310,7 +310,7 @@ ADD_MANAGED_METHOD(Shader, Load0, void, (MonoString* path, unsigned int* outReso
 ADD_MANAGED_METHOD(Shader, Load1, void,
 	(MonoString* path, MonoString* vertexPath, MonoString* fragmentPath,
 		MonoString* computePath, MonoString* geometryPath,
-		unsigned int* outResourceID, void** outHandle), AquaEngine.Graphics)
+		uint64_t* outResourceID, void** outHandle), AquaEngine.Graphics)
 {
 	ShaderStageInfo stageInfo =
 	{
