+++ Engine/Source/Scripting/InternalCalls/Resources.cpp
@@ -6,7 +6,7 @@ using namespace std;
 using namespace AquaEngine;
 using namespace AquaEngine::Scripting;
 
-ADD_MANAGED_METHOD(Resource, Duplicate, unsigned int, (unsigned int originalID, MonoString* newPath))
+ADD_MANAGED_METHOD(Resource, Duplicate, uint64_t, (unsigned int originalID, MonoString* newPath))
 {
 	return Resource::IsValidResourceID(originalID) ?
 		Resource::Duplicate(originalID, mono_string_to_utf8(newPath)) : InvalidResourceID;
@@ -18,14 +18,14 @@ ADD_MANAGED_METHOD(Resource, GetPath, MonoString*, (unsigned int id))
 		(Resource::IsValidResourceID(id) ? Resource::GetPath(id) : string()).c_str());
 }
 
-ADD_MANAGED_METHOD(Resource, GetID, unsigned int, (MonoString* path))
+ADD_MANAGED_METHOD(Resource, GetID, uint64_t, (MonoString* path))
 { return Resource::GetID(mono_string_to_utf8(path)); }
 
-ADD_MANAGED_METHOD(Resource, Unload, void, (unsigned int resourceID))
+ADD_MANAGED_METHOD(Resource, Unload, void, (uint64_t resourceID))
 { Resource::Unload(resourceID); }
 
-ADD_MANAGED_METHOD(Resource, GetInstance, void*, (unsigned int resourceID))
+ADD_MANAGED_METHOD(Resource, GetInstance, void*, (uint64_t resourceID))
 { return Resource::IsValidResourceID(resourceID) ? Resource::Get(resourceID) : nullptr; }
 
-ADD_MANAGED_METHOD(Resource, Exists, bool, (unsigned int resourceID))
+ADD_MANAGED_METHOD(Resource, Exists, bool, (uint64_t resourceID))
 { return Resource::IsValidResourceID(resourceID); }
